.azurepipelines/requirements.txt
edk2-pytool-library
PyGithub